// plant_tracker/frontend/src/services/models/area-in.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Book Store API
 * Book Store API for renting books and notifying available/returned books in a store
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AreaIn
 */
export interface AreaIn {
    /**
     * 
     * @type {string}
     * @memberof AreaIn
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof AreaIn
     */
    'location_id': string;
}